Add a very similar offset encoder, that changes the long offsets
The long offsets are offset from the longest short offset. This makes
the look-back window longer by the length of a short offset. In theory
it should yield a slight improvement in compression opportunities,
being able to look back a little further with no encoding penalty. But
requires the look-back buffer to be larger.

A few minor changes