Incremental decompression: check offset is within history
Avoid information leak if decoding invalid compressed data. Previously
this was handled by clearing the history buffer to zero at the start.
But this implements a check on-the-fly instead.

This method is faster for small amounts of data (by saving a large
memset() at the start), but expected to be slower for large amounts of
data (due to extra check throughout decompression).